Move RoleContext to contexts directory and update imports
