+++ src/pages/admin/Users.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { supabase } from '../../lib/supabase';
-import { useRole } from '../../context/RoleContext';
+import { useRole } from '../../contexts/RoleContext';
 import { Users as UsersIcon, Search, Filter, Edit, Trash, UserPlus, CheckCircle, XCircle } from 'lucide-react';
 import { Button } from '../../components/ui/button';
 import { Navigate } from 'react-router-dom';
